Checkstyle and copyright fixes.
